refactor: move UpsertSettings and settings_client.go in to client_settings.go
